+++ sys/isrpipe/read_timeout/Makefile
@@ -1,0 +1,3 @@
+MODULE = isrpipe_read_timeout
+
+include $(RIOTBASE)/Makefile.base
